requirements.txt
flask
python-dotenv
pymongo
requests
openai
notion-client
python-telegram-bot==20.0
google-auth
google-auth-oauthlib
google-auth-httplib2
pytz
